Cleaned up code
Cleaned up the coding style of the multilinechart and the
AnimatedChartingThing(which is now known as AnimatedSingleLineChart),
refactored the ChartingThing to make the name a bit more sensible.
Fixed a few known issues in the MultiLineChart(which were inherited from
the erros that were previously fixed in the ChartingThing and
AnimatedSingleLineChart).